feat: restructure repo
- Add TypeScript typings
- Add ESM exports
- Overall improve package.json
- Add files array to package.json
