+++ esm/presentation.mjs
@@ -1,0 +1,3 @@
+import mod from "../src/presentation.js";
+
+export default mod;
