+++ src/slide.js
