+++ src/templates/diagram.js
