+++ src/templates/simple.js
