+++ src/templates/split.js
