+++ src/templates/terminal.js
